requirements.txt
https://github.com/thiagopena/geraldo/archive/master.zip
https://github.com/thiagopena/PySIGNFe/archive/master.zip
Django==4.0.*
dj-database-url
python-decouple
reportlab
python-magic